move setup logic to separate file
